==> Figure3/02_Freq_Changes_and_DEG_CMV_SF4/01A_SF4_scRNA_Assemble_and_Get_Counts.ipynb <==
import pandas as pd 
import numpy as np 
import scanpy as sc
import matplotlib.pyplot as plt
import concurrent.futures
import pickle
import warnings
from datetime import date
import hisepy
import os
import shutil
from concurrent.futures import ThreadPoolExecutor, as_completed,ProcessPoolExecutor
from tqdm import tqdm
import anndata
import gc
warnings.filterwarnings("ignore")
sc.settings.n_jobs = 60
print("Current working directory:", os.getcwd())

meta_data_SF4 = pd.read_csv("/home/jupyter/BRI_Figures_Final_V1/Dataset/SF4_meta_data-2024-05-05.csv")


# Read MetaData

meta_data_SF4 = pd.read_csv("/home/jupyter/BRI_Figures_Final_V1/Dataset/SF4_meta_data-2024-05-05.csv")
meta_data_SF4['ageAtEnrollment'] = meta_data_SF4['ageAtEnrollment'].replace('89+', '90')
meta_data_SF4['ageAtEnrollment'] = pd.to_numeric(meta_data_SF4['ageAtEnrollment'], errors='coerce')
meta_data_SF4 = meta_data_SF4.dropna(subset=['ageAtEnrollment'])
meta_data_SF4['Ages'] = meta_data_SF4['ageAtEnrollment']
meta_data_SF4 = meta_data_SF4.drop_duplicates(subset='sample.sampleKitGuid', keep='last')

# Assemble Year 1 Day 0

%%time
file_names= ['/home/jupyter/BRI_Figures_Final_V1/Dataset/scRNA/SF4/h5ad/'+x+".h5ad" for x in meta_data_SF4['pbmc_sample_id'].tolist()]
adata_list = []
with ThreadPoolExecutor(max_workers=60) as executor:
    future_to_file = {executor.submit(sc.read_h5ad, file_name): file_name for file_name in file_names}
    for future in tqdm(as_completed(future_to_file), total=len(file_names)):
        result = future.result()
        if result is not None:
            adata_list.append(result)

adata = anndata.concat(adata_list,join='outer')

adata=adata[~adata.obs['cmv_status'].isna()]

fitlered_gene=pd.DataFrame()
for i in adata.obs['celltypist_l3'].unique():
    print(i)
    adata_subset=adata[adata.obs['celltypist_l3']==i]
    sc.pp.filter_genes(adata_subset, min_cells=round(np.shape(adata_subset.X)[0]*0.1))
    gene_list=pd.DataFrame(list(adata_subset.var.index))
    gene_list.columns=['gene']
    gene_list['AIFI_L3']=i
    fitlered_gene=pd.concat([fitlered_gene,gene_list])

fitlered_gene.to_csv("fitlered_gene_SF4.csv")

# Get Counts

file_names= ['/home/jupyter/BRI_Figures_Final_V1/Dataset/scRNA/SF4/h5ad/'+x+".h5ad" for x in meta_data_SF4['pbmc_sample_id'].tolist()]

def get_counts(sample_id):
    adata = sc.read_h5ad('/home/jupyter/BRI_Figures_Final_V1/Dataset/scRNA/SF4/h5ad/'+sample_id+".h5ad", backed='r')
    counts_df = pd.DataFrame(adata.obs["celltypist_l3"].value_counts()).reset_index()
    counts_df['pbmc_sample_id'] = sample_id
    return counts_df

with concurrent.futures.ProcessPoolExecutor(max_workers=60) as executor:
    results = list(tqdm(executor.map(get_counts, meta_data_SF4['pbmc_sample_id']), total=len(meta_data_SF4['pbmc_sample_id'])))

final_df = pd.concat(results, ignore_index=True)
final_df.to_csv("SF4_AIFI_L3_Counts.csv")
