==> Figure3/02_Freq_Changes_and_DEG_CMV_SF4/01C_SF4_Freq_Comparsion.ipynb <==
suppressPackageStartupMessages(library(plyr))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
library(parallel)
library(reshape2)
library(ggplot2)
library(patchwork)
library(ggpubr)
library(rstatix)
source('../../helper_function/helper_function_IHA.r')

meta_data=read.csv("../../Dataset/SF4_meta_data-2024-05-05.csv")
meta_data$ageAtEnrollment[meta_data$ageAtEnrollment == '89+'] <- '90'
meta_data$ageAtEnrollment<-as.numeric(meta_data$ageAtEnrollment)
meta_data<-meta_data[!is.na(meta_data$ageAtEnrollment),]
meta_data$Ages<-meta_data$ageAtEnrollment
#meta_data <- meta_data[!duplicated(meta_data$sample.sampleKitGuid, fromLast = TRUE), ]

table(meta_data$lab.revisionHistory.dataHistory.CMV.Ab.Screen.Result..CMVRES..newValue)

count_data<-read.csv("SF4_AIFI_L3_Counts.csv")

count_data <- count_data %>%
  dplyr::group_by(pbmc_sample_id) %>%                       
  dplyr::mutate(percentage = (count / sum(count)) * 100) %>% 
  dplyr::group_by(pbmc_sample_id) %>%     
  dplyr::mutate(percentage_clr = clr_transform((percentage))) %>% 
  dplyr::ungroup() %>% as.data.frame(.)

head(count_data)

meta_data<-meta_data %>% filter(lab.revisionHistory.dataHistory.CMV.Ab.Screen.Result..CMVRES..newValue%in% c("Negative",'Positive'))

meta_data$CMV<-meta_data$lab.revisionHistory.dataHistory.CMV.Ab.Screen.Result..CMVRES..newValue

count_data<-count_data %>% filter(pbmc_sample_id %in% meta_data$pbmc_sample_id)

count_data<-left_join(count_data,meta_data,by=c("pbmc_sample_id"))

count_data<-count_data %>% group_by(sample.sampleKitGuid) %>%
  dplyr::mutate(percentage = (count / sum(count)) * 100) %>% 
  ungroup()%>% 
  group_by(sample.sampleKitGuid) %>%     
  dplyr::mutate(AIFI_L3_clr = clr_transform((percentage))) %>%
  dplyr::ungroup()

log2fc<-count_data %>%
  select(percentage,AIFI_L3_clr, celltypist_l3, cohort.cohortGuid, subject.biologicalSex, CMV) %>%
  group_by(celltypist_l3, CMV) %>%
  summarise(mean_AIFI_L3_clr = median(AIFI_L3_clr, na.rm = TRUE)) %>%
  spread(CMV, mean_AIFI_L3_clr) %>%
  mutate(log2fc = log2(`Positive` / `Negative`)) %>%
  mutate(delta_change = `Positive` - `Negative`) 

STAT<-count_data %>%
  select(AIFI_L3_clr,celltypist_l3, cohort.cohortGuid,subject.biologicalSex,CMV) %>%
  group_by(celltypist_l3)  %>%
  wilcox_test(AIFI_L3_clr ~ CMV
             ) %>%
  adjust_pvalue(method='BH') %>%
  add_significance() %>% as.data.frame()

STAT<-left_join(STAT,log2fc)

colnames(STAT)[1]<-"AIFI_L3"

# Get DEG Number

DEG_res<-read.csv("Deseq2_Result_SF4.csv")

DEG_res_signi<-DEG_res %>% filter(padj<0.05,abs(log2FoldChange)>0.1)

DEG_counts<-as.data.frame(table(DEG_res_signi$contrast,DEG_res_signi$celltype))

colnames(DEG_counts)<-c("factor","AIFI_L3","DEG_Counts")

# Combine STAT


STAT<-left_join(STAT,DEG_counts )

STAT<-STAT%>%
  mutate(adjP_degs_combined_value = if_else(DEG_Counts != 0,
                                            -log10(p.adj)  * DEG_Counts * 0.1,
                                            -log10(p.adj)  * 0.1))

write.csv(STAT,"DEG_counts_and_Frequency_Changes.csv")
