==> Figure3/02_Freq_Changes_and_DEG_CMV_SF4/04_Count_DEG.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(parallel)

# Y1D90

df_BR1 <- read.csv("Deseq2_Result_BR1_Y1D90_CMVpos_vs_CMVneg.csv")
df_BR2 <- read.csv("Deseq2_Result_BR2_Y1D90_CMVpos_vs_CMVneg.csv")

df_BR1_filtered <- df_BR1 %>% filter(abs(log2FoldChange) > 0.1, padj < 0.05)
df_BR2_filtered <- df_BR2 %>% filter(abs(log2FoldChange) > 0.1, padj < 0.05)

BR1_DEG_counts <- df_BR1_filtered %>% count(celltype)
BR2_DEG_counts <- df_BR2_filtered %>% count(celltype)

combined_counts <- full_join(BR1_DEG_counts, BR2_DEG_counts, by = "celltype", suffix = c("_BR1", "_BR2"))

common_DEGs <- inner_join(df_BR1_filtered, df_BR2_filtered, by = c("gene", "celltype"))

common_DEG_counts <- common_DEGs %>% count(celltype)

combined_counts <- left_join(combined_counts, common_DEG_counts, by = "celltype")
combined_counts <- combined_counts %>% rename(Common_DEGs = n)

combined_counts[is.na(combined_counts)] <- 0

colnames(combined_counts)<-c("celltype",'BR1_counts','BR2_counts','Common_counts')

write.csv(combined_counts,'DEG_Counts_Y1D90.csv')

# Y2D0

df_BR1 <- read.csv("Deseq2_Result_BR1_Y2D0_CMVpos_vs_CMVneg.csv")
df_BR2 <- read.csv("Deseq2_Result_BR2_Y2D0_CMVpos_vs_CMVneg.csv")

df_BR1_filtered <- df_BR1 %>% filter(abs(log2FoldChange) > 0.1, padj < 0.05)
df_BR2_filtered <- df_BR2 %>% filter(abs(log2FoldChange) > 0.1, padj < 0.05)

BR1_DEG_counts <- df_BR1_filtered %>% count(celltype)
BR2_DEG_counts <- df_BR2_filtered %>% count(celltype)

combined_counts <- full_join(BR1_DEG_counts, BR2_DEG_counts, by = "celltype", suffix = c("_BR1", "_BR2"))

common_DEGs <- inner_join(df_BR1_filtered, df_BR2_filtered, by = c("gene", "celltype"))

common_DEG_counts <- common_DEGs %>% count(celltype)

combined_counts <- left_join(combined_counts, common_DEG_counts, by = "celltype")
combined_counts <- combined_counts %>% rename(Common_DEGs = n)

combined_counts[is.na(combined_counts)] <- 0

colnames(combined_counts)<-c("celltype",'BR1_counts','BR2_counts','Common_counts')

write.csv(combined_counts,'DEG_Counts_Y2D0.csv')
